Fetch users to display usrename properly
